Fixed animation does not work
Hi! @dimitrisraptis96 
Thanks for your great works!

I wanted to use eva icon with React, but I was in trouble because I could not find an example even if I looked for it.
So, this library was very helpful. 

There was only one problem, animation did not work.
I tried to correct it, so I want you to check this PR. Thanks!